Add and edit User tests

+++ src/test/java/site/nomoreparties/stellarburgers/user/UserCreateTest.java
@@ -5,6 +5,7 @@ import io.qameta.allure.Feature;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 
@@ -14,13 +15,17 @@ public class UserCreateTest {
 
     private final UserClient client = new UserClient();
     private final UserAssertions check = new UserAssertions();
+    private User user;
+    private String accessToken;
 
-    String accessToken;
+    @Before
+    public void createUser() {
+        user = User.random();
+    }
 
     @Test
     @DisplayName("Успешное создание уникального пользователя")
     public void userSuccessCreationTest() {
-        var user = User.random();
         ValidatableResponse createResponse = client.createUser(user);
         check.createdSuccessfully(createResponse, user);
         accessToken = client.getUserAccessToken(createResponse);
@@ -29,7 +34,6 @@ public class UserCreateTest {
     @Test
     @DisplayName("Ошибка при создании пользователя, который уже существует")
     public void userAlreadyExistsCreationTest() {
-        var user = User.random();
         ValidatableResponse createFirstResponse = client.createUser(user);
         accessToken = client.getUserAccessToken(createFirstResponse);
         ValidatableResponse createSecondResponse = client.createUser(user);
@@ -38,8 +42,9 @@ public class UserCreateTest {
 
     @After
     public void deleteUser() {
-        if (!accessToken.isEmpty())
+        if (accessToken != null && !accessToken.isEmpty()) {
             client.deleteUser(accessToken);
+        }
     }
