+++ src/test/java/site/nomoreparties/stellarburgers/user/UserLoginTest.java
@@ -5,6 +5,7 @@ import io.qameta.allure.Feature;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 
@@ -14,29 +15,27 @@ public class UserLoginTest {
 
     private final UserClient client = new UserClient();
     private final UserAssertions check = new UserAssertions();
+    private User user;
+    private String accessToken;
 
-    String accessToken;
+    @Before
+    public void createUser() {
+        user = User.random();
+        var createResponse = client.createUser(user);
+        accessToken = client.getUserAccessToken(createResponse);
+    }
 
     @Test
     @DisplayName("Успешная авторизация пользователя")
     public void userSuccessAuthTest() {
-        var user = User.random();
-        var createResponse = client.createUser(user);
-        accessToken = client.getUserAccessToken(createResponse);
-
         var userCredentials = UserCredentials.fromUser(user);
         ValidatableResponse loginResponse = client.loginUser(userCredentials);
         check.loginSuccessfully(loginResponse, user);
-
     }
 
     @Test
     @DisplayName("Логин с некорректным логином")
     public void userWrongLoginAuthTest() {
-        var user = User.random();
-        var createResponse = client.createUser(user);
-        accessToken = client.getUserAccessToken(createResponse);
-
         var userCredentials = new UserCredentials("wrongEmail@email.com", user.getPassword());
         ValidatableResponse loginResponse = client.loginUser(userCredentials);
         check.loginWrongEmailPasswordError(loginResponse);
@@ -45,10 +44,6 @@ public class UserLoginTest {
     @Test
     @DisplayName("Логин с некорректным паролем")
     public void userWrongPassAuthTest() {
-        var user = User.random();
-        var createResponse = client.createUser(user);
-        accessToken = client.getUserAccessToken(createResponse);
-
         var userCredentials = new UserCredentials(user.getEmail(), "wrongPassword");
         ValidatableResponse loginResponse = client.loginUser(userCredentials);
         check.loginWrongEmailPasswordError(loginResponse);
@@ -57,10 +52,6 @@ public class UserLoginTest {
     @Test
     @DisplayName("Логин с некорректным логином/паролем")
     public void userWrongCredentialsAuthTest() {
-        var user = User.random();
-        var createResponse = client.createUser(user);
-        accessToken = client.getUserAccessToken(createResponse);
-
         var userCredentials = new UserCredentials("wrongEmail", "wrongPassword");
         ValidatableResponse loginResponse = client.loginUser(userCredentials);
         check.loginWrongEmailPasswordError(loginResponse);
@@ -68,8 +59,9 @@ public class UserLoginTest {
 
     @After
     public void deleteUser() {
-        if (!accessToken.isEmpty())
+        if (accessToken != null && !accessToken.isEmpty()) {
             client.deleteUser(accessToken);
+        }
     }
 
 }
